Moved spigot related messaging into new file.
